entrego el dato del sensor de direccion en rads

+++ lib/SensorStatus/SensorStatus.h
@@ -6,7 +6,7 @@ class SensorStatus{
         SensorStatus(int pinclk_dir, int pindata_dir, int nbits_dir, int sincsim_dir);  //hay q ir agregando las fuciones
 
         /*int sincsim_dir es solo para el sensor simulado */
-        int get_dir();
+        float get_dir();
 
     private:
         int _pinclk_dir;
